check for existence of static variation element before renaming it
Upgrading Nosto to the latest version caused the checkout to no longer want to render.

```
Message: Unable to process binding "afterRender: function(){return reloadRecommendations }"
Message: Cannot read property 'classList' of null
    at UiClass.reloadRecommendations (variation-tagging.js:56)
    at init (after-render.js:20)
    at knockout.js:3358
    at Object.ignore (knockout.js:1470)
    at knockout.js:3357
    at Object.arrayForEach (knockout.js:159)
    at applyBindingsToNodeInternal (knockout.js:3343)
    at applyBindingsToNodeAndDescendantsInternal (knockout.js:3223)
    at applyBindingsToDescendantsInternal (knockout.js:3205)
    at Object.ko.applyBindingsToDescendants (knockout.js:3421)
```

According to the comment "Remove the static variation if it exists - it should not". So just check if element exists before trying to rename it.